Separated two phase landau polynomial in Base and Regressor version
