+++ apal/__init__.py
@@ -4,5 +4,6 @@ from apal.cahn_hilliard import CahnHilliard
 from apal.landau_poly_fit_constraints import PeakPosition
 from apal.landau_poly_fit_constraints import StraightLineSaddle
 from apal.landau_poly_fit_constraints import InteriorMinima
-from apal.landau_polynomial import TwoPhaseLandauPolynomial
+from apal.landau_polynomial import TwoPhaseLandauPolynomialBase
+from apal.kernel_regressor_two_phase import TwoPhasePolynomialRegressor
 from apal.khachaturyan import Khachaturyan
